requirements.txt
fastapi
uvicorn
pynacl
